corners_live/requirements.txt
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.1
pytz==2024.1
